requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# We'll add these later for ML/Data
# pandas==2.1.3
# scikit-learn==1.3.2
# textblob==0.17.1